Added clear number of plays button in Options + implemented grid creation for game screen.
A button to reset the number of times the game is played has been added to the options screen along with a save button for user to confirm changes to settings. All necessary elements have been implemented so that game screen is populated with grid respective of user chosen options.